+++ src/apps/exactly/common/exactly.deposit.token-fetcher.ts
@@ -1,14 +1,12 @@
 import { type BigNumber, constants } from 'ethers';
 
-import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
 import type { GetDataPropsParams, GetTokenPropsParams } from '~position/template/app-token.template.types';
 
 import type { ExactlyMarketDefinition } from '../common/exactly.definitions-resolver';
 import { type ExactlyMarketProps, ExactlyTokenFetcher } from '../common/exactly.token-fetcher';
 import type { Market } from '../contracts';
 
-@PositionTemplate()
-export class ExactlyDepositFetcher extends ExactlyTokenFetcher {
+export abstract class ExactlyDepositFetcher extends ExactlyTokenFetcher {
   groupLabel = 'Variable Deposit';
 
   getSupply({ definition }: GetTokenPropsParams<Market, ExactlyMarketProps, ExactlyMarketDefinition>) {